refactor: :recycle: Button을 prop으로 받도록 수정
